requirements.txt
numpy==2.0.2
numba==0.60.0
beautifulsoup4
dem_stitcher
fire
folium
geojson
geopandas
ipyleaflet
ipywidgets
matplotlib
notebook
opencv-python-headless
pandas
pdal
Pillow
pyproj
rasterio
requests
rpcm
scikit-image
scipy
shapely
srtm4
staticmap
tqdm
xmltodict
